FatFs R0.13c

+++ samples/lpc23xx/main.c
@@ -1,21 +1,20 @@
 /*----------------------------------------------------------------------*/
-/* FAT file system sample project for FatFs            (C)ChaN, 2013    */
+/* FAT file system sample project for FatFs            (C)ChaN, 2010    */
 /*----------------------------------------------------------------------*/
 
 #include <string.h>
-#include "integer.h"
 #include "interrupt.h"
 #include "rtc23xx.h"
 #include "uart23xx.h"
 #include "disp.h"
 #include "filer.h"
 #include "xprintf.h"
-#include "diskio.h"
 #include "ff.h"
+#include "diskio.h"
 
 
 
-DWORD AccSize;				/* Work register for fs command */
+QWORD AccSize;				/* Work register for fs command */
 WORD AccFiles, AccDirs;
 FILINFO Finfo;
 
@@ -574,7 +573,7 @@ int main (void)
 				xprintf("...");
 				res = scan_files(ptr);
 				if (res) { put_rc(res); break; }
-				xprintf("\r%u files, %lu bytes.\n%u folders.\n"
+				xprintf("\r%u files, %llu bytes.\n%u folders.\n"
 						"%lu KiB total disk space.\n%lu KiB available.\n",
 						AccFiles, AccSize, AccDirs,
 						(fs->n_fatent - 2) * (fs->csize / 2), (DWORD)p1 * (fs->csize / 2)
